requirements.txt
tkinter
pillow
matplotlib.pyplot